this is the initial design proposal
for the content declarations page. SWT does not support
wrapped table headers (https://stackoverflow.com/a/53534738).
Tooltips could be an alternative.

Cell editing is not really possible because of the different
types of data in the table (compononents -> materials
-> substances; value ranges).

So a dialog and table  actions like
`add component|material|substance`, `edit` & `delete` are the
way to go...