this is maybe confusing, but for each question we have
two objects: one that is stored in the configuration
and one that is stored in the Q-Meta data of the data
set.

instead of the question IDs we pass directly the
question from the configuration into the respective
section. configuration questions do not have an
answer but the questions in the data set have.

it would be cleaner two have different kind of
classes for this, but ...

next, we have to think about where the question
configuration is stored, maybe just as a file that
can be imported