#27 The column width is now set in the following way:
Say that the table has a width of 100%. We give 10% to the UUID column
and 10%
to the version column. The view where the validation messages are shown
is a
generic status view where sometimes names are long and the status
messages are
short or the other way around. Thus, we try to assign the column widths
for the
name and status message in relation to the content of these columns,
with a
minimum (maximum) width of 15% (65%) for both columns (we assume that
the
character width is in average the same for bot columns):

```
totalLength = maxNameLenght + maxMessageLength
nameWidth = maxNameLength / totalLength
nameWidth = min(.65, max(.15, nameWidth))
messageWidth = maxMessageLength / totalLength
messageWidth = min(.65, max(.15, messageWidth))
```